+++ packages/crayons-core/src/components/form/form.e2e.ts
@@ -157,6 +157,17 @@ describe('fw-form', () => {
           parent_id: null,
           choices: [],
         },
+        {
+          id: 'f319f86f-1b6a-49cb-b4b6-cf4873674595',
+          name: 'profile_pic',
+          label: 'Profile picture',
+          type: 'FILES',
+          position: 7,
+          required: true,
+          placeholder: '',
+          hint: 'Accepts only PNG',
+          multiple: true,
+        },
       ],
     },
   };
@@ -485,7 +496,7 @@ describe('fw-form', () => {
     const element = await page.find('fw-form');
     expect(
       element.shadowRoot.querySelectorAll('fw-form-control').length
-    ).toEqual(6);
+    ).toEqual(7);
   });
 
   it('should render correct crayons controls based on the field name', async () => {
@@ -496,6 +507,7 @@ describe('fw-form', () => {
       DROPDOWN: 'fw-select',
       NUMBER: 'fw-input',
       DECIMAL: 'fw-input',
+      FILES: 'fw-file-uploader-2',
     };
     const page = await newE2EPage();
 
@@ -503,7 +515,7 @@ describe('fw-form', () => {
 
     await page.$eval(
       'fw-form',
-      (elm: any, { formSchema }) => {
+      (elm: any, { formSchema }: any) => {
         elm.formSchema = formSchema;
       },
       props
@@ -520,6 +532,7 @@ describe('fw-form', () => {
       formControl[3].shadowRoot.querySelector('fw-radio-group');
     const dropdown = formControl[4].shadowRoot.querySelector('fw-select');
     const decimal = formControl[5].shadowRoot.querySelector('fw-input');
+    const files = formControl[6].shadowRoot.querySelector('fw-file-uploader-2');
 
     expect(input).not.toBeNull();
     expect(props.formSchema.fields[0].id).toEqual(input.id);
@@ -547,6 +560,10 @@ describe('fw-form', () => {
     expect(map[props.formSchema.fields[5].type]).toEqual(
       decimal.tagName.toLowerCase()
     );
+    expect(files).not.toBeNull();
+    expect(map[props.formSchema.fields[6].type]).toEqual(
+      files.tagName.toLowerCase()
+    );
   });
 
   it('should return entered values upon form submit', async () => {
@@ -556,7 +573,7 @@ describe('fw-form', () => {
 
     await page.$eval(
       'fw-form',
-      (elm: any, { formSchema }) => {
+      (elm: any, { formSchema }: any) => {
         elm.formSchema = formSchema;
         elm.initialValues = {
           first_name: 'Test',
@@ -565,6 +582,24 @@ describe('fw-form', () => {
           pincode: 123345,
           order_status: 'closed',
           amount_paid: 10,
+          profile_pic: [
+            {
+              file: new File(
+                [
+                  new Blob(
+                    new Uint8Array([
+                      137, 80, 78, 71, 13, 10, 26, 10, 0, 0, 0, 13, 73, 72, 68,
+                      82, 0, 0, 0, 8, 0, 0, 0, 8, 8, 2, 0, 0, 0, 75, 109, 41,
+                      220, 0, 0, 0, 34, 73, 68, 65, 84, 8, 215, 99, 120, 173,
+                      168, 135, 21, 49,
+                    ])
+                  ),
+                ],
+                'file1.png',
+                { type: 'png', lastModified: Date.now() }
+              ),
+            },
+          ],
         };
       },
       props
@@ -580,6 +615,8 @@ describe('fw-form', () => {
     expect(result.values['pincode']).toEqual(123345);
     expect(result.values['order_status']).toEqual('closed');
     expect(result.values['amount_paid']).toEqual(10);
+    // TODO: fileList is being received as object, need to understand reason.
+    expect(Object.keys(result.values['profile_pic']).length).toEqual(1);
   });
 
   it('should set errors if required field is passed as empty string', async () => {
@@ -589,7 +626,7 @@ describe('fw-form', () => {
 
     await page.$eval(
       'fw-form',
-      (elm: any, { formSchema }) => {
+      (elm: any, { formSchema }: any) => {
         elm.formSchema = formSchema;
         elm.initialValues = {
           first_name: 'name',
@@ -597,6 +634,7 @@ describe('fw-form', () => {
           gender: '',
           order_status: '',
           amount_paid: '',
+          profile_pic: [],
         };
       },
       props
@@ -613,6 +651,7 @@ describe('fw-form', () => {
     expect(result.errors['gender']).toEqual('Gender is required');
     expect(result.errors['order_status']).toEqual('Order Status is required');
     expect(result.errors['amount_paid']).toEqual('Amount Paid is required');
+    expect(result.errors['profile_pic']).toEqual('Profile picture is required');
   });
 
   it('Should return number for decimal and number field type', async () => {
@@ -900,7 +939,7 @@ describe('fw-form', () => {
 
     await page.$eval(
       'fw-form',
-      (elm: any, { formSchema }) => {
+      (elm: any, { formSchema }: any) => {
         elm.formSchema = formSchema;
         elm.initialValues = {
           pincode: 123345,
@@ -929,7 +968,7 @@ describe('fw-form', () => {
 
     await page.$eval(
       'fw-form',
-      (elm: any, { formSchema }) => {
+      (elm: any, { formSchema }: any) => {
         elm.formSchema = formSchema;
         elm.initialValues = {
           pincode: 123345,
@@ -948,7 +987,7 @@ describe('fw-form', () => {
     await page.waitForChanges();
     expect(
       element.shadowRoot.querySelectorAll('fw-form-control').length
-    ).toEqual(6);
+    ).toEqual(7);
   });
 
   it('should disabled the fields for which the editable property is set to false in form schema', async () => {
@@ -958,7 +997,7 @@ describe('fw-form', () => {
 
     await page.$eval(
       'fw-form',
-      (elm: any, { formSchema }) => {
+      (elm: any, { formSchema }: any) => {
         elm.formSchema = formSchema;
       },
       props